cache types during normalization
